onrequirements.txt
streamlit
onnxruntime
numpy
Pillow
